requirements.txt
flask==2.1.0
flask-mysqldb==1.0.1
python-dotenv==0.20.0
werkzeug==2.1.1
apscheduler==3.9.1
reportlab==4.2.5
gunicorn==20.1.0